double

+++ src/com/PracticalTraining/Task3.java
@@ -1,15 +1,15 @@
 package com.PracticalTraining;
 
 public class Task3 {
-    public static int add(int num1, int num2) {
+    public static double add(double num1, double num2) {
         return num1 + num2;
     }
 
-    public static int subtract(int num1, int num2) {
+    public static double subtract(double num1, double num2) {
         return num1 - num2;
     }
 
-    public static int multiply(int num1, int num2) {
+    public static double multiply(double num1, double num2) {
         return num1 * num2;
     }
 
@@ -24,12 +24,12 @@ public class Task3 {
 
 class CalculatorTest {
     public static void main(String[] args) {
-        int num1 = 10;
-        int num2 = 5;
+        double num1 = 10;
+        double num2 = 5;
 
-        int sum = Task3.add(num1, num2);
-        int difference = Task3.subtract(num1, num2);
-        int product = Task3.multiply(num1, num2);
+        double sum = Task3.add(num1, num2);
+        double difference = Task3.subtract(num1, num2);
+        double product = Task3.multiply(num1, num2);
         double quotient = Task3.divide(num1, num2);
 
         System.out.println("和: " + sum);
